producer/requirements.txt
psycopg2-binary
datetime
faker 
kafka-python 
python-json-logger